rename placeholder to integrations
