+++ public/components/integrations/components/added_integration_table.tsx
