+++ public/components/integrations/components/available_integration_table.tsx
