+++ public/components/integrations/components/integration_assets_panel.tsx
@@ -1,7 +1,7 @@
 import { EuiInMemoryTable, EuiLink, EuiPanel, EuiSpacer, EuiTableFieldDataColumnType, EuiText } from '@elastic/eui';
 import { FILTER_OPTIONS } from '../../../../common/constants/explorer';
 import React from 'react';
-import { PanelTitle } from '../../../../public/components/trace_analytics/components/common/helper_functions';
+import { PanelTitle } from '../../trace_analytics/components/common/helper_functions';
 import _ from 'lodash';
 
 export function IntegrationAssets(props: any) {
