+++ public/components/integrations/components/integration_card.tsx
