+++ public/components/integrations/components/integration_details_panel.tsx
@@ -12,7 +12,7 @@ import {
   EuiPageContentHeaderSection,
 } from '@elastic/eui';
 import React from 'react';
-import { PanelTitle } from '../../../../public/components/trace_analytics/components/common/helper_functions';
+import { PanelTitle } from '../../trace_analytics/components/common/helper_functions';
 
 export function IntegrationDetails(props: any) {
   let screenshots = undefined;
