+++ public/components/integrations/components/integration_overview_panel.tsx
